% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chlmap.R
\name{cdommap}
\alias{cdommap}
\title{Create cdom maps}
\usage{
cdommap(
  yearmon,
  subgroup = NA,
  stream.qa = T,
  remove.flags = TRUE,
  fdir = getOption("fdir")
)
}
\arguments{
\item{yearmon}{numeric date in yyyymm format}

\item{subgroup}{subgroup of grab sample stations if creating separate regressions based on area}

\item{stream.qa}{logical Use QA'd streaming data?}

\item{remove.flags}{logical Use QA'd grab data?}

\item{fdir}{file.path to data folder}
}
\value{
An extracted chlorophyll surface and an updated FullDataSet file
}
\description{
Create a cdom concentration surface using streaming data and regression against extracted cdom
}
\details{
A new interpolation is run after calculating an extracted cdom for all streaming observations. Calculated values that exceed the maximum observed grab sample concentration are discarded.
}
\author{
Jemma Stachelek and Kira Allen
}
