% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grabclean.R
\name{grabclean}
\alias{grabclean}
\title{Cleaning grab data}
\usage{
grabclean(yearmon, tofile = FALSE, fdir = getOption("fdir"))
}
\arguments{
\item{yearmon}{numeric survey date in yyyymm format}

\item{tofile}{logical save output to file}

\item{fdir}{character file path to local data directory}
}
\description{
format grab sample data
}
\details{
If streaming data does not exist for a particular data/time pull averages for the previous minute (if data exists).
}
\examples{
\dontrun{
res <- grabclean(yearmon = 202412, tofile = FALSE)
}
}
