% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QAdash.R
\name{QAdash}
\alias{QAdash}
\title{Creates interactive dashboard for data QA}
\usage{
QAdash(yearmon, original.data = FALSE, fdir = getOption("fdir"))
}
\arguments{
\item{yearmon}{numeric designation of survey date formatted as yyyymm}

\item{original.data}{specify whether to read in original streamcleaned data (TRUE) or already QAd data (FALSE)}

\item{fdir}{character file path to local data directory}
}
\description{
Creates interactive dashboard for data QA
}
\examples{
\dontrun{
QAdash(yearmon=202510, original.data=TRUE)
}
}
