% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QA.R
\name{QAflags}
\alias{QAflags}
\title{Create plots that flag data based on values or change}
\usage{
QAflags(
  yearmon,
  param,
  bad_min,
  bad_max,
  sus_min,
  sus_max,
  step_threshold,
  min_time = 17,
  max_time = 9,
  plots,
  original.data = FALSE,
  fdir = getOption("fdir")
)
}
\arguments{
\item{yearmon}{numeric designation of survey date formatted as yyyymm}

\item{param}{name of parameter to be examined}

\item{bad_min}{values less than this are flagged as bad}

\item{bad_max}{values more than this are flagged as bad}

\item{sus_min}{values less than this are flagged as suspect}

\item{sus_max}{values more than this are flagged as suspect}

\item{step_threshold}{values flagged if change between susequent measurements are greater than this threshold}

\item{min_time}{data after this time (24 hour clock) are not plotted}

\item{max_time}{data before this time (24 hour clock) are not plotted}

\item{plots}{user specifies which plots to create (flag points, flag map, step flag points, step flag map), can specify 2 at a time}

\item{original.data}{specify whether to read in original streamcleaned data (TRUE) or already QAd data (FALSE)}

\item{fdir}{character file path to local data directory}
}
\description{
Create plots that flag data based on values or change
}
\examples{
\dontrun{
d <- QAflags(yearmon=202510, param="tal.pc.rfu", bad_min=0, bad_max=10, sus_min=0.01, sus_max=4, step_threshold=2,plots=c("flag points", "step flag points"), original.data=F)
}
}
