% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QA.R
\name{replace_data}
\alias{replace_data}
\title{Replace bad data values}
\usage{
replace_data(
  df = df,
  yearmon,
  value_var,
  replace_val = NA,
  min_value = NULL,
  max_value = NULL,
  start_date = NULL,
  end_date = NULL,
  date_var = "timestamp",
  flag = NULL,
  flag_var = "flag",
  step_flag = NULL,
  step_flag_var = "step_flag",
  logic = "or",
  save = TRUE,
  fdir = getOption("fdir")
)
}
\arguments{
\item{df}{dataframe (if specifying values based on QA flags, the dataframe must contain the columns with those flags)}

\item{yearmon}{numeric designation of survey date formatted as yyyymm}

\item{value_var}{name of parameter of interest}

\item{min_value}{replace values less than this}

\item{max_value}{replace values more than this}

\item{start_date}{replace values occuring after this datetime (must be used with end_date)}

\item{end_date}{replace values before this datetime (must be used with start_date)}

\item{date_var}{the date/time column that start_date and end_date reference (default is timestamp column created in QAflags function)}

\item{flag}{selects values with a given flag}

\item{flag_var}{column name that contains flags}

\item{step_flag}{selects values with given step flag}

\item{step_flag_var}{column name that contains step flags}

\item{logic}{whether to follow "and" or "or"conditions when selecting values}

\item{save}{whether to save output to QA datasets folder}

\item{fdir}{character file path to local data directory}

\item{replace_value}{what to replace selected values with (default is to set selected values to NA)}
}
\description{
Replace bad data values based on user-specified criteria
}
\examples{
\dontrun{
df <- replace_data(df=data, yearmon=202510, value_var="tal.pc.rfu", flag="Bad", step_flag=TRUE, logic="and", save=T)
df <- replace_data(df=data, yearmon=202510, value_var="tal.pc.rfu", min_value=0, replace_val=0, save=T)
dat <- replace_data(df=data, yearmon=202510, value_var="tal.pc.rfu", start_date="2025-10-21 12:57:42", end_date="2025-10-21 13:13:13", replace_val=20, save=F)
}
}
