% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Streaminterp.R
\name{streaminterp}
\alias{streaminterp}
\title{streaminterp}
\usage{
streaminterp(
  dt,
  paramlist,
  yearmon,
  trim_rstack = TRUE,
  trim_negative = TRUE,
  costrasname = "CostRas_Optimize_Clip.tif",
  tname = NA,
  vname = NA,
  missprop = 0.16,
  range = 3750,
  fdir = getOption("fdir")
)
}
\arguments{
\item{dt}{input data frame}

\item{paramlist}{list of parameters (dt column names) to interpolate}

\item{yearmon}{a file path used to extract basename}

\item{trim_rstack}{logical trim the raster stack by the convex hull of training points?}

\item{trim_negative}{logical set negative values to zero?}

\item{costrasname}{character file.path to cost raster}

\item{tname}{file.path location to save training dataset}

\item{vname}{file.path location to save validation dataset}

\item{missprop}{numeric proportion of missing data allowed. Variables with a greater proportion of missing data will be dropped.}

\item{range}{interpolation range}

\item{fdir}{character file path to local data directory}
}
\description{
streaminterp
}
\examples{
\dontrun{
dt <- streamget(yearmon = 201513, qa = TRUE)
streaminterp(dt, paramlist = c("sal"), yearmon = 201513)
}
}
